This works but not what i wanted
because i have to set the binding context manually in xaml causing the
view model’s constructor to be called again. Also i have to override
the range navigator’s series in the code behind.

Because of this there are “two” instances of the PageModel (view model)
and my app will be updating one at runtime but it won’t cause the graph
to update.